// apps/command_agent_version_2/app/lib/model/innerEnums.ts
export enum InputTextKeys {
    TEXT = "text",
}

export enum OutputTextKeys {
    TEXT = "text",
}

export enum PromptKeys {
    TEXT = "text",
    MODEL = "model",
    TEMPERATURE = "temperature",
    MAX_TOKENS = "maxTokens",
}

export enum AgentKeys {
    DESCRIPTION = "description",
    QUERY = "query",
    PERSONALITY = "personality",
    INSTRUCTIONS = "agentInstructions",
    MODEL = "model",
    CHARGE_TYPE = "modelChargeType",
    OUTPUT_TYPE = "outputType",
    TEMPERATURE = "temperature",
    MAX_TOKENS = "maxTokens",
    TOOLS = "tools",
}